Moved general/music commands to their own file
